Update README.md
Hai there, I'm Ganesh! 👋

Welcome to my GitHub profile! I'm a passionate software engineer with a Bachelor's degree in Engineering (BE) and a fervent enthusiasm for creating innovative solutions. My journey in the world of technology has been fueled by curiosity and a drive to learn, grow, and collaborate with like-minded individuals.